requirements.txt
aiohttp==3.6.3
async-timeout==3.0.1
attrs==20.2.0
discord.py==1.5.1
loguru==0.5.3
multidict==4.7.6
SQLAlchemy==1.3.20
yarl==1.5.1
